requirements.txt
appdirs==1.4.3
biopython==1.71
certifi==2018.4.16
chardet==3.0.4
ConfigArgParse==0.13.0
datrie==0.7.1
decorator==4.3.0
docutils==0.14
idna==2.7
Jinja2==2.10
jsonschema==2.6.0
llvmlite==0.23.2
MarkupSafe==1.0
networkx==2.1
numba==0.38.1
numpy==1.14.5
pandas==0.23.1
pkg-resources==0.0.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
ratelimiter==1.2.0.post0
requests==2.19.1
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
snakemake==5.1.5
tabview==1.4.3
tqdm==4.23.4
umap-learn==0.2.5
urllib3==1.23
wrapt==1.10.11
